+++ tests/test_image.py
@@ -5,14 +5,14 @@ import numpy as np
 
 import torch
 
-import HyperXAI as hx
+import meteors as mt
 
 class TestImageMethods(unittest.TestCase):
     
     def test_image(seld):
         sample = torch.tensor([[[0]]])
         wavelengths = [0]
-        image = hx.HyperImage(image=sample, wavelengths=wavelengths, binary_mask="artificial")
+        image = mt.Image(image=sample, wavelengths=wavelengths, binary_mask="artificial")
         
     
     def test_wavelengths(self):
@@ -20,14 +20,14 @@ class TestImageMethods(unittest.TestCase):
         with self.assertRaises(ValueError):
                 wavelengths = [0, 1]
 
-                image = hx.HyperImage(image=sample, wavelengths=wavelengths)
+                image = mt.Image(image=sample, wavelengths=wavelengths)
                     
     
     
     def test_artificial_mask(self):
         sample = torch.tensor([[[0]]])
         wavelengths = [0]
-        image = hx.HyperImage(image=sample, wavelengths=wavelengths, binary_mask="artificial")
+        image = mt.Image(image=sample, wavelengths=wavelengths, binary_mask="artificial")
         self.assertEqual(image.binary_mask, torch.tensor([[[0]]]))
         
     def test_incorrect_shape_mask(self):
@@ -36,11 +36,11 @@ class TestImageMethods(unittest.TestCase):
         with self.assertRaises(ValueError):
             binary_mask = torch.tensor([[[0, 0]]])
 
-            image = hx.HyperImage(image=sample, wavelengths=wavelengths, binary_mask=binary_mask)
+            image = mt.Image(image=sample, wavelengths=wavelengths, binary_mask=binary_mask)
         
         with self.assertRaises(ValueError):
             binary_mask = "very bad mask"
-            image = hx.HyperImage(image=sample, wavelengths=wavelengths, binary_mask=binary_mask)
+            image = mt.Image(image=sample, wavelengths=wavelengths, binary_mask=binary_mask)
             
 
     def test_rgb_image(self):